requirements.txt
astroid==2.9.3
autopep8==1.6.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
DateTime==4.4
ecdsa==0.17.0
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
platformdirs==2.5.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.0.0
pylint==2.12.2
pytz==2021.3
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
typing_extensions==4.1.1
uuid==1.30
Werkzeug==2.0.3
wrapt==1.13.3
zope.interface==5.4.0
